TASKS/requirements.txt
Bottleneck @ file:///tmp/build/80754af9/bottleneck_1607575130224/work
certifi==2021.10.8
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
et-xmlfile==1.1.0
fonttools==4.25.0
kiwisolver @ file:///tmp/build/80754af9/kiwisolver_1612281846643/work
lxml==4.6.4
matplotlib @ file:///tmp/build/80754af9/matplotlib-suite_1638289681807/work
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186066731/work
mkl-service==2.4.0
munkres==1.1.4
numexpr @ file:///tmp/build/80754af9/numexpr_1618856529730/work
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1634095651905/work
olefile @ file:///Users/ktietz/demo/mc3/conda-bld/olefile_1629805411829/work
openpyxl==3.0.9
packaging @ file:///tmp/build/80754af9/packaging_1637314298585/work
pandas==1.3.4
Pillow==8.4.0
pyparsing @ file:///tmp/build/80754af9/pyparsing_1635766073266/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python-docx==0.8.11
pytz==2021.3
sip==4.19.13
six @ file:///tmp/build/80754af9/six_1623709665295/work
tornado @ file:///tmp/build/80754af9/tornado_1606942317143/work
XlsxWriter==3.0.2
